Add ParOr and On constructs + refactor
Each coroutine still have a unique Entity bind to it, but now all the
data is stored in the Executor (might change later). ParOr allocate each
coroutines on the heap, but might figure out a better way later. There
is still a lot of bookkeeping in the Executor that might be removed in a
future commit. Also add tests.