Make Executor Send + Sync -> now is a Resource
That also means that we can depend on the official bevy repo now,
and not on a fork. That also means that coroutines need to be
Send + Sync now, which adds some limitation when sharing values
from the outside, but some convenient APIs will be build for
that later on.